requirements.txt
pyspark==3.5.1
Flask
Flask-Cors
pandas
numpy
anthropic
python-dotenv
openpyxl
pyarrow
xlsxwriter
chardet
reportlab
# DebtGuardian Framework Dependencies
ollama
pydantic>=2.0.0
guardrails-ai
GitPython